Always allow the cf re-bind functionality
On number of occasions, we've faced the leaked S3 access keys. When this
happens, AWS is removing our set of statements on the user policy. This
causes the broker, unable to perform unbind action, as there is no
statement to delete.

We want to relax that functionality, so that users are always empowered
to re-bind the service causing the access token rotation, without
extreme measures and the GOV.UK PaaS Support intervention.

This change involves very werid practice, of returning both error and
the result. But then checking the error kind and deciding whether or not
to kill the action, or simply log the error and carry on. Its only
purpose, is to leave the function as pure as possible, and not have a
dependency on the logger.